app/core/studio/requirements.txt
google-cloud-storage>=3.1.1
langchain>=0.3.25
langchain-core>=0.3.65
langchain-mistralai>=0.2.10
langgraph>=0.4.8
openai>=1.88.0
pandas>=2.3.0
python-dotenv>=1.1.0
scikit-learn>=1.7.0
snowflake-connector-python[pandas]>=3.15.0
trustcall>=0.0.39
ipython